+++ source/Public/Set-PSModulePath.ps1
@@ -17,6 +17,15 @@
         If set the PSModulePath will be changed machine wide. If not set, only
         the current session will be changed.
 
+    .PARAMETER FromTarget
+        The target environment variable to copy the value from.
+
+    .PARAMETER ToTarget
+        The target environment variable to set the value to.
+
+    .PARAMETER PassThru
+        If specified, returns the set value.
+
     .EXAMPLE
         Set-PSModulePath -Path '<Path 1>;<Path 2>'
 
@@ -28,6 +37,12 @@
 
         Sets the machine environment variable `PSModulePath` to the specified path
         or paths (separated with semi-colons).
+
+    .EXAMPLE
+        Set-PSModulePath -FromTarget 'MAchine' -ToTarget 'User'
+
+        Copies the value of the machine environment variable `PSModulePath` to the
+        user environment variable `PSModulePath`.
 #>
 function Set-PSModulePath
 {
@@ -36,25 +51,64 @@ function Set-PSModulePath
         '',
         Justification = 'ShouldProcess is not supported in DSC resources.'
     )]
-    [CmdletBinding()]
+    [CmdletBinding(DefaultParameterSetName = 'Default')]
     param
     (
-        [Parameter(Mandatory = $true)]
+        [Parameter(Mandatory = $true, ParameterSetName = 'Default')]
         [ValidateNotNullOrEmpty()]
         [System.String]
         $Path,
 
+        [Parameter(ParameterSetName = 'Default')]
+        [System.Management.Automation.SwitchParameter]
+        $Machine,
+
+        [Parameter(Mandatory = $true, ParameterSetName = 'TargetParameters')]
+        [ValidateSet('Session', 'User', 'Machine')]
+        [System.String]
+        $FromTarget,
+
+        [Parameter(Mandatory = $true, ParameterSetName = 'TargetParameters')]
+        [ValidateSet('Session', 'User', 'Machine')]
+        [System.String]
+        $ToTarget,
+
         [Parameter()]
         [System.Management.Automation.SwitchParameter]
-        $Machine
+        $PassThru
     )
 
-    if ($Machine.IsPresent)
+    if ($PSCmdlet.ParameterSetName -eq 'Default')
     {
-        [System.Environment]::SetEnvironmentVariable('PSModulePath', $Path, [System.EnvironmentVariableTarget]::Machine)
+        if ($Machine.IsPresent)
+        {
+            [System.Environment]::SetEnvironmentVariable('PSModulePath', $Path, [System.EnvironmentVariableTarget]::Machine)
+        }
+        else
+        {
+            $env:PSModulePath = $Path
+        }
     }
-    else
+    elseif ($PSCmdlet.ParameterSetName -eq 'TargetParameters')
     {
-        $env:PSModulePath = $Path
+        $Path = Get-EnvironmentVariable -Name 'PSModulePath' -FromTarget $FromTarget
+
+        switch ($ToTarget)
+        {
+            'Session'
+            {
+                [System.Environment]::SetEnvironmentVariable('PSModulePath', $Path)
+            }
+
+            default
+            {
+                [System.Environment]::SetEnvironmentVariable('PSModulePath', $Path, [System.EnvironmentVariableTarget]::$ToTarget)
+            }
+        }
+    }
+
+    if ($PassThru.IsPresent)
+    {
+        return $Path
     }
 }